src/targetFunction/requirements.txt
azure-functions
numpy
matplotlib